Make Void Acid break blocks; remove Sculk Extractor; add Memory Block Comparator; update manual, creative tab order, textures and add more recipes; Experience Interface supports Experience Fluid tag, reduced scan interval and removed transfer limit; fix Redstone logic. 使虚空酸可破坏方块；删除幽匿萃取体；添加记忆方块比较器；更新手册、创造物品栏顺序、材质并新增更多配方；经验接口支持经验流体标签，扫描间隔降低且去除转移上限；修正红石逻辑。

+++ src/main/java/dev/anvilcraft/pigsplus/fluid/VoidAcidFluid.java
@@ -13,26 +13,34 @@ import net.minecraft.world.level.material.Fluid;
 import net.minecraft.world.level.material.FluidState;
 import net.neoforged.neoforge.fluids.BaseFlowingFluid;
 
+import static dev.anvilcraft.pigsplus.AnvilCraftPigsPlus.CONFIG;
+
 public abstract class VoidAcidFluid extends BaseFlowingFluid {
     protected VoidAcidFluid(Properties properties) {
         super(properties);
     }
 
     @Override
+    protected boolean isRandomlyTicking() {
+        return true;
+    }
+
+    @Override
     public void randomTick(Level level, BlockPos pos, FluidState state, RandomSource random) {
-        // TODO: 不知道为什么不触发
         super.randomTick(level, pos, state, random);
+
+        if (level.isClientSide()) return;
+        if (!CONFIG.voidAcidDestroy) return;
 
         for (Direction direction : Direction.values()) {
             BlockPos targetPos = pos.relative(direction);
             BlockState targetState = level.getBlockState(targetPos);
 
             if (targetState.is(AddonBlockTags.VOID_ACID_IMMUNE)) continue;
-
             float hardness = targetState.getDestroySpeed(level, targetPos);
             if (hardness <= 0.0f) continue;
 
-            float probability = 5f / (hardness + 10.0f);
+            float probability = 1f / (hardness + 2.0f);
             if (random.nextFloat() < probability) {
                 level.destroyBlock(targetPos, false);
                 if (random.nextFloat() < 0.1f) {
